+++ lib/down/errors.rb
@@ -20,7 +20,7 @@ module Down
   class ResponseError < NotFound
     attr_reader :response
 
-    def initialize(message, response:)
+    def initialize(message, response: nil)
       super(message)
       @response = response
     end